[Phone] Deny access phone number.

+++ patcher/src/main/java/li/lingfeng/ltsystem/patcher/Patcher.java
@@ -44,6 +44,7 @@ public class Patcher {
     private static final Map<String, String> PACKAGE_PATH_MAP = new HashMap<String, String>() {{
         put("java.lang", "/libcore/ojluni/src/main/java/");
         put("java.util", "/libcore/ojluni/src/main/java/");
+        put("android.telephony", "/frameworks/base/telephony/java/");
         put("android.media", "/frameworks/base/media/java/");
         put("com.android.server", "/frameworks/base/services/core/java/");
         put("com.android.internal.telephony", "/frameworks/opt/telephony/src/java/");
@@ -207,10 +208,6 @@ public class Patcher {
                         FileUtils.writeStringToFile(file, content);
                     }
                     return;
-                }
-            } else if (token.getCategory() == JavaToken.Category.SEPARATOR) {
-                if (token.getText().equals("{")) {
-                    throw new Exception("Can't find method name position.");
                 }
             }
         }
@@ -427,6 +424,9 @@ public class Patcher {
     }
 
     private void copyAdditionalFiles() throws Throwable {
+        if (SIMULATE) {
+            return;
+        }
         Logger.i("Copy additional files.");
         String srcDir = "./additional_files";
         FileUtils.copyDirectory(new File(srcDir), new File(Config.ANDROID_SOURCE_DIR), false);
